// packages/convex/src/component/_generated/api.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * To regenerate, run `npx convex dev`.
 * @module
 */

import { anyApi, componentsGeneric } from 'convex/server';

import type * as lib from '../lib.js';
import type * as shared from '../shared.js';
import type * as utils from '../utils.js';
import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from 'convex/server';

const fullApi: ApiFromModules<{
  lib: typeof lib;
  shared: typeof shared;
  utils: typeof utils;
}> = anyApi as any;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, 'public'>
> = anyApi as any;

/**
 * A utility for referencing Convex functions in your app's internal API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = internal.myModule.myFunction;
 * ```
 */
export const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, 'internal'>
> = anyApi as any;

export const components = componentsGeneric() as unknown as {
  rateLimiter: import('@convex-dev/rate-limiter/_generated/component.js').ComponentApi<'rateLimiter'>;
  transactionalPool: import('@convex-dev/workpool/_generated/component.js').ComponentApi<'transactionalPool'>;
  bulkPool: import('@convex-dev/workpool/_generated/component.js').ComponentApi<'bulkPool'>;
  callbackPool: import('@convex-dev/workpool/_generated/component.js').ComponentApi<'callbackPool'>;
};